Add secondary service manager support

+++ whtree/modules/platform/js/bootstrap/servicemanager.ts
@@ -1,15 +1,35 @@
 /* For now: experimental service runner to replace webhare.cpp and decentral service managers
    Invoke using wh run mod::platform/js/bootstrap/servicemanager.ts
+
+   When debugging us, it may be useful to run a second instance for easy restarting. To do this:
+   - Start your primary instance with a different service name:
+     wh console --name platform:altsm --exclude "webhare_testsuite_temp:*"
+   - Start a secondary instance
+     wh run mod::platform/js/bootstrap/servicemanager.ts --secondary -v --include "webhare_testsuite_temp:*"
+
+  If you're stuck with a lot of stray processes on OSX an effective way to kill them all is:
+  kill $(ps eww -ax|grep ' WEBHARE_NOINSTALLATIONINFO=1'|cut -d' ' -f1)
 */
 
 import * as fs from 'node:fs';
 import { debugFlags } from "@webhare/env/src/envbackend";
-import { backendConfig } from "@webhare/services/src/services";
+import { backendConfig, resolveResource } from "@webhare/services/src/services";
 import { storeDiskFile } from "@webhare/system-tools/src/fs";
 import * as child_process from "child_process";
-import { createDeferred, sleep } from "@webhare/std";
+import { createDeferred, sleep, wildcardsToRegExp } from "@webhare/std";
 import { getCompileServerOrigin, getRescueOrigin } from "@mod-system/js/internal/configuration";
 import { RotatingLogFile } from "../logging/rotatinglogfile";
+import runBackendService from '@mod-system/js/internal/webhareservice';
+import { program } from 'commander'; //https://www.npmjs.com/package/commander
+import { getAllModuleYAMLs } from '@webhare/services/src/moduledefparser';
+
+program.name("servicemanager")
+  .option("-s, --secondary", "Mark us as a secondary service manager")
+  .option("-v, --verbose", "Verbose logging (also set by 'startup' debug flag)")
+  .option("--name [servicename]", "Name for the backend service to manage us", "platform:servicemanager")
+  .option("--include <mask>", "Only manage services that match this mask", "")
+  .option("--exclude <mask>", "Do not manage services that match this mask", "")
+  .parse();
 
 enum Stages { Bootup, StartupScript, Active, Terminating, ShuttingDown }
 let currentstage = Stages.Bootup;
@@ -21,6 +41,10 @@ const MinimumRunTime = 60000;
 const MaxStartupDelay = 60000;
 const MaxLineLength = 512;
 const earlywebserver = process.env.WEBHARE_WEBSERVER == "node";
+const isSecondaryManager: boolean = program.opts().secondary;
+const verbose = program.opts().verbose || debugFlags.startup;
+
+let keepAlive: NodeJS.Timeout | null = null;
 let shuttingdown = false;
 let logfile;
 
@@ -49,7 +73,7 @@ const stagetitles: Record<Stages, string> = {
   [Stages.ShuttingDown]: "Shutting down bridge and database"
 };
 
-const expectedServices: Record<string, ServiceDefinition> = {
+const defaultServices: Record<string, ServiceDefinition> = {
   "platform:whmanager": {
     cmd: ["whmanager"],
     startIn: Stages.Bootup,
@@ -98,6 +122,8 @@ const expectedServices: Record<string, ServiceDefinition> = {
     keepAlive: true
   }
 };
+
+const expectedServices = new Map<string, ServiceDefinition>();
 
 class ProcessManager {
   readonly name;
@@ -221,6 +247,8 @@ class ProcessManager {
       return;
 
     this.toldToStop = true;
+    if (verbose)
+      log(`Stopping service ${this.displayName}`);
     this.process!.kill(this.service.stopSignal ?? "SIGTERM");
 
     const timeout = this.service.stopTimeout ?? DefaultTimeout;
@@ -240,20 +268,26 @@ class ProcessManager {
 
 const processes = new Set<ProcessManager>;
 
+/// Move to a new stage
 async function startStage(stage: Stages): Promise<void> {
-  const subpromises = [];
-  process.title = `webhare: ${stagetitles[stage]} `;
-  currentstage = stage;
-  if (debugFlags.startup)
+  if (verbose)
     log(`Starting stage: ${stagetitles[stage]} `);
+  currentstage = stage;
+  return await updateForCurrentStage();
+}
+
+/// Actually apply the current stage. Also used when configurationchanges
+async function updateForCurrentStage(): Promise<void> {
+  const subpromises = [];
+  process.title = `webhare: ${stagetitles[currentstage]} `;
 
   for (const process of [...processes]) {
-    if ((process.service.stopIn ?? DefaultShutdownStage) === stage)
+    if ((process.service.stopIn ?? DefaultShutdownStage) === currentstage || !expectedServices.has(process.name))
       subpromises.push(process.stop());
   }
 
-  for (const [name, service] of Object.entries(expectedServices)) {
-    if (service.startIn === stage && !service.current) {
+  for (const [name, service] of expectedServices.entries()) {
+    if (service.startIn === currentstage && !service.current) {
       const proc = new ProcessManager(name, service);
       if (service.waitForCompletion)
         subpromises.push(proc.stopDefer.promise); //TODO should we have a timeout? (but what do you do if it hits? terminate? move to next stage?)
@@ -297,25 +331,91 @@ function unlinkServicestateFiles() {
   signal(SIGTTOU, SIG_IGN);
 */
 
+class ServiceManagerClient {
+  getWebHareState() {
+    return {
+      stage: stagetitles[currentstage],
+      availableServices: [...expectedServices.entries()].map(([name, service]) => ({
+        name
+      }))
+    };
+  }
+  async reload() {
+    await loadServiceList(); //we block this so the service will be visible in the next getWebhareState from this client
+    updateForCurrentStage(); //I don't think we need to block clients on startup of services ? they can wait themslevs
+  }
+}
+
+export async function gatherManagedServices(): Promise<Record<string, ServiceDefinition>> {
+  const services: Record<string, ServiceDefinition> = {};
+
+  for (const mod of await getAllModuleYAMLs()) {
+    if (mod.managedServices)
+      for (const [name, servicedef] of Object.entries(mod.managedServices)) {
+        if (servicedef?.script) {
+          const cmd = ["wh", "run", resolveResource(mod.baseResourcePath, servicedef.script)];
+          services[`${mod.module}:${name}`] = {
+            cmd,
+            startIn: Stages.Active,
+            keepAlive: true
+          };
+        }
+      }
+  }
+
+  return services;
+}
+
+async function loadServiceList() {
+  const include = program.opts().include ? new RegExp(wildcardsToRegExp(program.opts().include)) : null;
+  const exclude = program.opts().exclude ? new RegExp(wildcardsToRegExp(program.opts().exclude)) : null;
+  const allservices = Object.entries({ ...await gatherManagedServices(), ...defaultServices });
+
+  const removeServices = new Set(expectedServices.keys());
+
+  for (const [name, servicedef] of allservices) {
+    if ((include && !include.test(name)) || (exclude && exclude.test(name)) || (isSecondaryManager && !include))
+      continue;
+
+    expectedServices.set(name, servicedef);
+    removeServices.delete(name);
+  }
+
+  for (const service of removeServices) {
+    expectedServices.get(service)!.keepAlive = false;
+    expectedServices.delete(service);
+  }
+}
+
 async function main() {
   if (!backendConfig.dataroot) {
     console.error("Cannot start WebHare. Data root not set");
     return 1;
   }
 
+  keepAlive = setInterval(() => { }, 1000 * 60 * 60 * 24); //keep us alive
+
+  //TODO check if webhare isn't already running when not started with --secondary
+
+  await loadServiceList();
   fs.mkdirSync(backendConfig.dataroot + "log", { recursive: true });
-  logfile = new RotatingLogFile(backendConfig.dataroot + "log/servicemanager", { stdout: true });
+  logfile = new RotatingLogFile(isSecondaryManager ? null : backendConfig.dataroot + "log/servicemanager", { stdout: true });
 
   const showversion = process.env.WEBHARE_DISPLAYBUILDINFO ?? backendConfig.buildinfo.version ?? "unknown";
   log(`Starting WebHare ${showversion} in ${backendConfig.dataroot} at ${getRescueOrigin()}`);
 
   //remove old servicestate files
-  unlinkServicestateFiles();
-
-  storeDiskFile(backendConfig.dataroot + ".webhare.pid", process.pid.toString() + "\n", { overwrite: true });
+  if (!isSecondaryManager) {
+    unlinkServicestateFiles();
+    storeDiskFile(backendConfig.dataroot + ".webhare.pid", process.pid.toString() + "\n", { overwrite: true });
+  }
 
   await startStage(Stages.Bootup);
-  await waitForCompileServer();
+  if (!isSecondaryManager)
+    await waitForCompileServer();
+
+  //FIXME do we need to wait for the bridge to be ready? do we auto reregister when te bridge comes bakck?
+  runBackendService(program.opts().name, () => new ServiceManagerClient, { autoRestart: false, dropListenerReference: true });
 
   if (!shuttingdown)
     await startStage(Stages.StartupScript);
@@ -331,15 +431,22 @@ async function shutdownSignal(signal: NodeJS.Signals) {
 }
 
 async function shutdown() {
+  if (shuttingdown)
+    return;
+  if (keepAlive)
+    clearTimeout(keepAlive);
+
   shuttingdown = true;
   await startStage(Stages.Terminating);
   await startStage(Stages.ShuttingDown);
-  try {
-    fs.unlinkSync(backendConfig.dataroot + ".webhare.pid");
-  } catch (e) {
-    console.error("Failed to remove webhare.pid file", e);
-  }
 
+  if (!isSecondaryManager) {
+    try {
+      fs.unlinkSync(backendConfig.dataroot + ".webhare.pid");
+    } catch (e) {
+      console.error("Failed to remove webhare.pid file", e);
+    }
+  }
 }
 
 process.on("SIGINT", shutdownSignal);
@@ -349,4 +456,7 @@ process.on("uncaughtException", (err, origin) => {
   shutdown();
   log(`Uncaught exception`);
 });
+
 main().then(exitcode => { process.exitCode = exitcode; }, e => console.error(e));
+
+export type { ServiceManagerClient };
